Renaming files since v2 no longer works
